Move messages-shared.ts to /query-server
